ref(dashboards): Move hasUnresolvedTraceMetric to its own file
