+++ static/app/views/dashboards/utils/getWidgetConfigError.tsx
@@ -1,9 +1,9 @@
 import {t} from 'sentry/locale';
-import {explodeField, isEquation} from 'sentry/utils/discover/fields';
 import {DisplayType, WidgetType, type Widget} from 'sentry/views/dashboards/types';
 import {usesTimeSeriesData} from 'sentry/views/dashboards/utils';
 import {extractTraceMetricFromColumn} from 'sentry/views/dashboards/widgetBuilder/utils/buildTraceMetricAggregate';
 import {getSelectedAggregate} from 'sentry/views/dashboards/widgetBuilder/utils/getSelectedAggregate';
+import {hasUnresolvedTraceMetric} from 'sentry/views/dashboards/widgetBuilder/utils/hasUnresolvedTraceMetric';
 import {doesMetricSupportHeatMapVisualization} from 'sentry/views/explore/metrics/constants';
 
 /**
@@ -37,18 +37,4 @@ export function getWidgetConfigError(widget: Widget): string | undefined {
   }
 
   return undefined;
-}
-
-/**
- * Whether any of a trace-metric widget's aggregates fail to resolve to a metric
- * (`fn(value, name, type, unit)`). This is the same predicate the metric selector's
- * auto-select uses, so it also answers "is a metric about to be filled in?".
- * Equations carry no metric tuple and reference base aggregates validated on their
- * own, so they're skipped. Callers must confirm the trace-metrics dataset first.
- */
-export function hasUnresolvedTraceMetric(widget: Widget): boolean {
-  return widget.queries
-    .flatMap(query => query.aggregates)
-    .filter(aggregate => !isEquation(aggregate))
-    .some(aggregate => !extractTraceMetricFromColumn(explodeField({field: aggregate})));
 }
